+++ react/Form.tsx
@@ -1,33 +1,64 @@
 import React, { FC } from 'react'
+import { useQuery } from 'react-apollo'
+import { useIntl, defineMessages } from 'react-intl'
+import { Alert } from 'vtex.styleguide'
 
-import { ObjectRenderer } from './components/ObjectRenderer'
+import { ObjectMapper } from './components/Object'
 import FormSubmit from './FormSubmit'
-import documentSchemaV2 from './graphql/getSchema.graphql'
-import { useQuery } from './hooks/mockUseQuery'
-import { FormRenderer } from './components/FormRenderer'
+import documentPublicSchema from './graphql/getSchema.graphql'
+import { FormHandler } from './components/FormHandler'
 import { FormProps } from './typings/FormProps'
 
+const messages = defineMessages({
+  loadingForm: {
+    id: 'store/form.schema.loading',
+    defaultMessage: '',
+  },
+  errorLoadingForm: {
+    id: 'store/form.schema.error',
+    defaultMessage: '',
+  },
+})
+
 const Form: FC<FormProps> = props => {
-  const { data } = useQuery(documentSchemaV2, {
+  const { entity, schema, children } = props
+  const intl = useIntl()
+  const { data, loading, error } = useQuery(documentPublicSchema, {
     variables: {
-      dataEntity: 'Test',
-      schema: 'person',
+      dataEntity: entity,
+      schema: schema,
     },
   })
 
-  if (!React.Children.count(props.children)) {
+  if (loading) {
     return (
-      <FormRenderer schema={data} formProps={props}>
-        <ObjectRenderer pointer="#" />
+      <Alert type="warning">{intl.formatMessage(messages.loadingForm)}</Alert>
+    )
+  }
+
+  if (error) {
+    return (
+      <Alert type="error">
+        {intl.formatMessage(messages.errorLoadingForm)}
+      </Alert>
+    )
+  }
+
+  const schemaDocument = data.documentPublicSchema.schema
+
+  if (!React.Children.count(children)) {
+    return (
+      <FormHandler schema={schemaDocument} formProps={props}>
+        <ObjectMapper pointer="#" />
         <FormSubmit label="Submit" />
-      </FormRenderer>
+      </FormHandler>
     )
   }
 
   return (
-    <FormRenderer schema={data} formProps={props}>
-      {props.children}
-    </FormRenderer>
+    <FormHandler schema={schemaDocument} formProps={props}>
+      {children}
+    </FormHandler>
   )
 }