+++ react/FormInput-dropdown.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-import { DropdownInput } from './components/DropdownRenderer'
-import { BaseInputProps } from './typings/InputProps'
-
-export default function FormInput(props: BaseInputProps) {
-  return <DropdownInput pointer={props.pointer} />
-}