+++ react/FormInputCheckbox.tsx
@@ -1,0 +1,3 @@
+import { CheckboxInput } from './components/Checkbox'
+
+export default CheckboxInput