+++ react/FormInputRadiogroup.tsx
@@ -1,0 +1,3 @@
+import { RadioGroupInput } from './components/RadioGroup'
+
+export default RadioGroupInput