+++ react/components/Checkbox.tsx
@@ -1,25 +1,30 @@
 import React, { FC } from 'react'
-import { Checkbox } from 'vtex.styleguide'
+import { Checkbox as StyleguideCheckbox } from 'vtex.styleguide'
 import { UseCheckboxReturnType, useCheckbox } from 'react-hook-form-jsonschema'
 
-export const CheckboxInput: FC<{ pointer: string }> = props => {
+import { BaseInputProps } from '../typings/InputProps'
+
+export const CheckboxInput: FC<BaseInputProps> = props => {
   const checkboxObject = useCheckbox(props.pointer)
-  return <CheckboxRenderer checkboxObject={checkboxObject} />
+  return <Checkbox checkboxObject={checkboxObject} label={props.label} />
 }
 
-export const CheckboxRenderer: FC<{
+export const Checkbox: FC<{
   checkboxObject: UseCheckboxReturnType
+  label?: string
 }> = props => {
   const checkboxObject = props.checkboxObject
+  const subSchema = checkboxObject.getObject()
+  const label = props.label ?? subSchema.title ?? checkboxObject.name
 
   if (checkboxObject.isSingle) {
     const checked = checkboxObject.formContext.watch(checkboxObject.pointer)
 
     return (
       <>
-        <Checkbox
+        <StyleguideCheckbox
           {...checkboxObject.getItemInputProps(0)}
-          label={checkboxObject.getObject().title}
+          label={label}
           required={checkboxObject.isRequired}
           value="true"
           {...(checked ? { checked: true } : { checked: false })}
@@ -41,10 +46,10 @@ export const CheckboxRenderer: FC<{
           `${checkboxObject.pointer}[${index}]`
         )
         return (
-          <Checkbox
+          <StyleguideCheckbox
             {...checkboxObject.getItemInputProps(index)}
             key={`${value}${index}`}
-            label={checkboxObject.getObject().title}
+            label={label}
             required={checkboxObject.isRequired}
             value={value}
             {...(checked ? { checked: true } : { checked: false })}