+++ react/components/Dropdown.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useMemo } from 'react'
-import { Dropdown } from 'vtex.styleguide'
+import { Dropdown as StyleguideDropdown } from 'vtex.styleguide'
 import {
   Controller,
   UseSelectReturnType,
@@ -7,18 +7,23 @@ import {
 } from 'react-hook-form-jsonschema'
 
 import { useFormattedError } from '../hooks/useErrorMessage'
+import { BaseInputProps } from '../typings/InputProps'
 
-export const DropdownInput: FC<{ pointer: string }> = props => {
+export const DropdownInput: FC<BaseInputProps> = props => {
   const selectObject = useSelect(props.pointer)
-  return <DropdownRenderer selectObject={selectObject} />
+  return <Dropdown selectObject={selectObject} label={props.label} />
 }
 
-export const DropdownRenderer: FC<{
+export const Dropdown: FC<{
   selectObject: UseSelectReturnType
+  label?: string
 }> = props => {
   const selectObject = props.selectObject
   const error = selectObject.getError()
-  const title = selectObject.getObject().title
+
+  const subSchema = selectObject.getObject()
+  const label = props.label ?? subSchema.title ?? selectObject.name
+
   const items = selectObject.getItems()
   const options = useMemo(() => {
     return items.map(value => {
@@ -33,10 +38,10 @@ export const DropdownRenderer: FC<{
         control={selectObject.formContext.control}
         rules={selectObject.validator}
         as={
-          <Dropdown
-            name={title}
+          <StyleguideDropdown
+            name={label}
             multi={false}
-            label={title}
+            label={label}
             options={options}
             error={error ? true : false}
             errorMessage={useFormattedError(error)}