+++ react/typings/InputProps.ts
@@ -1,0 +1,20 @@
+import { UISchemaType } from 'react-hook-form-jsonschema'
+
+export interface BaseInputProps {
+  pointer: string
+  label?: string
+}
+
+export interface FormFieldGroupProps extends Omit<BaseInputProps, 'label'> {
+  uiSchema?: UISchemaType
+}
+
+export enum InputTypes {
+  input = 'input',
+  hidden = 'hidden',
+  password = 'password',
+}
+
+export interface FormRawInputProps extends BaseInputProps {
+  inputType?: InputTypes
+}